requirements.txt
pyusb
bitstring
numpy
matplotlib